requirements.txt
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.1.0
pywebview>=5.0.0
matplotlib>=3.8.0
openpyxl>=3.1.0
numpy>=1.26.0
pillow>=10.0.0